Move responseCache tests into separate file
