+++ test/index.test.js
@@ -6,58 +6,7 @@ import {
   responseCache,
   CacheMissError,
 } from '../src';
-import {
-  successfulResponse,
-  jsonResponse,
-  emptyResponse,
-  serverErrorResponse,
-} from './responses';
-
-function hangingPromise() {
-  return new Promise(() => {});
-}
-
-fetchMock.get('/test/hangs', hangingPromise());
-fetchMock.get('/test/hangs/2', hangingPromise());
-fetchMock.get(
-  '/test/succeeds',
-  () =>
-    new Promise(resolve => {
-      resolve(successfulResponse());
-    })
-);
-
-fetchMock.get(
-  '/test/succeeds/json',
-  () =>
-    new Promise(resolve => {
-      resolve(jsonResponse());
-    })
-);
-
-fetchMock.post(
-  '/test/succeeds/json',
-  () =>
-    new Promise(resolve => {
-      resolve(jsonResponse());
-    })
-);
-
-fetchMock.get(
-  '/test/succeeds/empty',
-  () =>
-    new Promise(resolve => {
-      resolve(emptyResponse());
-    })
-);
-
-fetchMock.get(
-  '/test/fails/internal-server-error',
-  () =>
-    new Promise(resolve => {
-      resolve(serverErrorResponse());
-    })
-);
+import { successfulResponse } from './responses';
 
 describe('duplicateRequests.isRequestInFlight', () => {
   test('renders false when it is not in flight', () => {
@@ -789,88 +738,5 @@ describe('cachePolicy', () => {
         done();
       }
     );
-  });
-});
-
-describe('responseCache.has', () => {
-  test('behaves as expected', done => {
-    expect(responseCache.has('test')).toBe(false);
-    bestfetch('/test/succeeds/json', {
-      requestKey: 'test',
-    }).then(res => {
-      expect(res).toEqual(
-        expect.objectContaining({
-          data: {
-            a: true,
-          },
-          status: 200,
-          statusText: 'OK',
-          ok: true,
-        })
-      );
-      expect(responseCache.has('test')).toBe(true);
-      done();
-    });
-  });
-});
-
-describe('responseCache.delete', () => {
-  test('behaves as expected', done => {
-    expect(responseCache.delete('test')).toBe(false);
-    bestfetch('/test/succeeds/json', {
-      requestKey: 'test',
-    }).then(res => {
-      expect(res).toEqual(
-        expect.objectContaining({
-          data: {
-            a: true,
-          },
-          status: 200,
-          statusText: 'OK',
-          ok: true,
-        })
-      );
-      expect(responseCache.delete('test')).toBe(true);
-      expect(responseCache.has('test')).toBe(false);
-      done();
-    });
-  });
-});
-
-describe('responseCache.get', () => {
-  test('behaves as expected', done => {
-    expect(responseCache.get('test')).toBeUndefined();
-    bestfetch('/test/succeeds/json', {
-      requestKey: 'test',
-    }).then(() => {
-      expect(responseCache.get('test').data).toEqual({ a: true });
-      done();
-    });
-  });
-});
-
-describe('responseCache.set', () => {
-  test('behaves as expected when cache has response', done => {
-    bestfetch('/test/succeeds/json', {
-      requestKey: 'test',
-    }).then(() => {
-      const response = responseCache.get('test');
-      expect(response.data).toEqual({ a: true });
-      responseCache.set(
-        'test',
-        Object.assign({}, response, {
-          data: { a: false },
-        })
-      );
-      expect(responseCache.get('test').data).toEqual({ a: false });
-      done();
-    });
-  });
-  test('behaves as expected when response has not been cached', () => {
-    const response = responseCache.get('test');
-    expect(response).toBeUndefined();
-    responseCache.set('test', { data: { a: false } });
-    const newResponse = responseCache.get('test');
-    expect(newResponse).toEqual({ data: { a: false } });
   });
 });
